+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Automatic expansion of standard Google error message types in [error details](https://cloud.google.com/apis/design/errors#error_details) by [@optiman](https://github.com/optiman)
+- Cyclic detection of protobuf messages, which previous resulted in a stack overflow crash
+
+### Fixed
+- Oneof value that has a message with no fields would cause the UI to become unresponsive
+
 ## [v0.3.2] - 2020-11-25
 
 ### Fixed